requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-HTTPAuth==4.7.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.6
wandb==0.15.2
requests==2.28.2
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
flask-restx==1.1.0
langchain==0.0.165
python-dotenv==1.0.0
streamlit==1.22.0
huggingface_hub